P0-f-2b B2-1: 角色落地链类型迁移 float→double (纯类型, 无值-bug)
Scope (B2-1, character landing chain TYPE MIGRATION only; NO value-bug/threshold changes):
- CharacterMechanics.cs: verticalVelocityBeforeLanding field+ctor float→double; drop (float)runtime.Vy at :198
- LF2Character.cs: HandleLandingEvent(float→double)
- LF2Entity.cs: ApplySharedCharacterDatLandingIfNeeded(float→double landedVy)
- LF2CharacterDamageStateResolver.cs: HandleLandingEvent + ApplyFrozenLanding float→double;
  :390 unbox 'is float'→'is double'(primary)+'is float'(defensive fallback) to avoid silent Runtime.Vy fallback
- NOT touched (deferred): value-bugs 1f/3f (B2-2), 5f/7f (B2-3); thresholds 11.0f/17f/9.0f/7f/-3.5f (safe as double-widened)

Validation:
- Unity compile: ZERO errors (production + test; test HandleLandingEvent(18f/5f) float widens implicitly)
- 60-tick seed 12345 rerun: NO DIVERGENCE, 91 gated fields (incl Vy) all green, 172 aligned/0-orphan
- Pure type migration; characters do not land in this scenario so no value change, no regression

Codex/architect MCP unavailable entire batch (findcg.com 404/504); direct-evidence plan per authority.

+++ Assets/NTSD/Scripts/Animation/Character/CharacterMechanics.cs
@@ -72,7 +72,7 @@ namespace NTSD.Animation
         public readonly float visualYOffset;
         public readonly BoundaryResolveMode boundaryMode;
         public readonly bool landed;
-        public readonly float verticalVelocityBeforeLanding;
+        public readonly double verticalVelocityBeforeLanding; // P0-f-2b B2-1: float→double (character landing Vy snapshot, no truncation)
 
         public MechanicsStepResult(
             bool grounded,
@@ -80,7 +80,7 @@ namespace NTSD.Animation
             float visualYOffset,
             BoundaryResolveMode boundaryMode,
             bool landed = false,
-            float verticalVelocityBeforeLanding = 0f)
+            double verticalVelocityBeforeLanding = 0.0) // P0-f-2b B2-1: float→double
         {
             this.grounded = grounded;
             this.groundPlanePos = groundPlanePos;
@@ -195,7 +195,7 @@ namespace NTSD.Animation
                 UnitFriction(runtime);
             }
 
-            float vyBeforeVerticalMove = (float)runtime.Vy;
+            double vyBeforeVerticalMove = runtime.Vy; // P0-f-2b B2-1: no (float) truncation — double landing Vy snapshot
             runtime.Y += runtime.Vy;
 
             bool landed = runtime.Y > 0.0001f;
